# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  version: 1.0.0
  description: |
    Contract-driven API for Team-based Task Management System.
    Roles supported:
      - MANAGER
      - MEMBER

servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: http://your-vps-ip:8000/api/v1
    description: VPS deployment

tags:
  - name: Health
  - name: Auth
  - name: Projects
  - name: Tasks

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check endpoint
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              full_name: "Aarav Mehta"
              email: "aarav.mehta@company.com"
              password: "StrongPassword@123"
              role: "MANAGER"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "409":
          description: Email already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "aarav.mehta@company.com"
              password: "StrongPassword@123"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /projects:
    post:
      tags: [Projects]
      summary: Create project (MANAGER only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
            example:
              name: "Internal CRM System"
              description: "CRM system for managing internal client workflows"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (Only managers allowed)

    get:
      tags: [Projects]
      summary: List projects for authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Unauthorized

  /projects/{project_id}/members:
    get:
      tags:
        - Projects
      summary: List project members
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMemberListResponse'

        "401":
          description: Not authorized
        "403":
          description: Forbidden
        "404":
          description: Project not found

  /projects/{project_id}/members:
    post:
      tags:
        - Projects
      summary: Add member to project
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMemberAdd"
            example:
              user_id: "user_abc123"

      responses:
        "201":
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMemberResponse"
        "401":
          description: Not authorized
        "403":
          description: Forbidden
        "404":
          description: Project or user not found

  /projects/{project_id}/members/{user_id}:
    delete:
      tags:
        - Projects
      summary: Remove member from project
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed successfully
        "403":
          description: Not authorized
        "404":
          description: Not found

  /projects/{project_id}/tasks:
    post:
      tags: [Tasks]
      summary: Create task under project (MANAGER only)
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
            example:
              title: "Design database schema"
              description: "Create normalized schema with relationships"
              status: "ACTIVE"
              assigned_to: 5
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Project not found

  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks (Filterable)
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, PENDING, COMPLETE]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Unauthorized

  /tasks/{task_id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Unauthorized
        "404":
          description: Task not found

    put:
      tags: [Tasks]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
            example:
              title: "Design optimized DB schema"
              description: "Add indexing strategy"
              status: "PENDING"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Task not found

    delete:
      tags: [Tasks]
      summary: Soft delete task
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task moved to trash
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Task not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [full_name, email, password, role]
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [MANAGER, MEMBER]

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: "#/components/schemas/UserResponse"

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [MANAGER, MEMBER]

    ProjectCreateRequest:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string

    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time

    ProjectMemberUserInfo:
      type: object
      properties:
        id:
          type: string
          example: "user_abc123"
        email:
          type: string
          format: email
          example: "member@example.com"
      required:
        - id
        - email

    ProjectMemberListResponse:
      type: object
      properties:
        id:
          type: string
          example: "pm_789xyz"
        project_id:
          type: string
          example: "proj_123abc"
        joined_at:
          type: string
          format: date-time
          example: "2026-02-18T12:34:56Z"
        user:
          $ref: '#/components/schemas/ProjectMemberUserInfo'
      required:
        - id
        - project_id
        - joined_at
        - user

    ProjectMemberAdd:
      type: object
      properties:
        user_id:
          type: string
      required:
        - user_id

    ProjectMemberResponse:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        user_id:
          type: string
        joined_at:
          type: string
          format: date-time

    TaskCreateRequest:
      type: object
      required: [title, description, status, assigned_to]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, PENDING, COMPLETE]
        assigned_to:
          type: integer

    TaskUpdateRequest:
      type: object
      required: [title, description, status]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, PENDING, COMPLETE]

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, PENDING, COMPLETE]
        assigned_to:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean